# docs/api/openapi.yml
openapi: 3.1.0
info:
  title: Award Monitoring & Tracking System API
  version: 1.0.0
  description: |
    RESTful API for enterprise award monitoring and tracking platform.
    
    ## Authentication
    The application is its own OAuth 2.1 / OpenID Connect provider. Browser clients obtain tokens through the
    authorization code flow with PKCE and send `Authorization: Bearer <access token>` on every `/api` request.
    
    ## Rate Limiting
    - Standard users: 1000 requests/hour
    - Admin users: 5000 requests/hour
    
    ## Response Formats
    All responses are JSON. Dates are ISO 8601 format (UTC).
    
    ## Error Handling
    Errors follow RFC 7807 Problem Details format.
    
  contact:
    name: API Support
    email: support@award-system.edu.ua
    url: https://award-system.edu.ua/docs
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.award-system.edu.ua/v1
    description: Production server
  - url: https://staging-api.award-system.edu.ua/v1
    description: Staging server
  - url: http://localhost:8080/api/v1
    description: Local development
...
paths:
...
  /users/me:
    get:
      tags: [Users]
      summary: Get current user profile
      security:
        - bearerAuth: []
      operationId: getCurrentUser
      responses:
        '200':
          description: User profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'

    patch:
      tags: [Users]
      summary: Update current user profile
      security:
        - bearerAuth: []
      operationId: updateCurrentUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdateRequest'
      responses:
        '200':
          description: Profile updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
  /users:
    get:
      tags: [Users]
      summary: User directory
      description: |
        Everybody for `user:read:all`; for `user:read:scope` only members of the organisations inside the
        caller's role scopes. Unverified accounts never appear. Sorted by last name, first name.
      security:
        - bearerAuth: []
      operationId: listUsers
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
        - name: organization
          in: query
          description: Limit to this organisation and its descendants; must lie inside the caller's scope
          schema:
            type: integer
            format: int64
        - name: role
          in: query
          description: Holders of the role today
          schema:
            $ref: '#/components/schemas/UserRole'
        - name: status
          in: query
          description: Without it, every status except `PENDING` and `DELETED`
          schema:
            $ref: '#/components/schemas/AccountStatus'
        - name: unconfirmed
          in: query
          description: Only users who have never held a role
          schema:
            type: boolean
            default: false
        - name: q
          in: query
          description: Case-insensitive fragment of the name or address; ignored under two characters
          schema:
            type: string
      responses:
        '200':
          description: One page; `size` is capped at 100
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserPage'
        '403':
          description: |
            `urn:awards:problem:access-denied` — the caller lacks `user:read:all`/`user:read:scope`, or
            `organization` is outside the caller's scope; the `detail` names what is missing
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
  /users/{id}:
    get:
      tags: [Users]
      summary: One user with the role history
      security:
        - bearerAuth: []
      operationId: getUser
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: The user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDetail'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          description: Unknown, unverified, or outside the caller's scope (the directory does not confirm foreign ids)
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'

  # ============== Reports ==============
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Access token issued by /oauth2/token
  parameters:
    PageParam:
      name: page
      in: query
      description: Page number (0-based)
      schema:
        type: integer
        default: 0
        minimum: 0
    SizeParam:
      name: size
      in: query
      description: Page size
      schema:
        type: integer
        default: 20
        minimum: 1
        maximum: 100
...
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: integer
          format: int64
        email:
          type: string
          format: email
        firstName:
          type: string
        lastName:
          type: string
        roles:
          type: array
          items:
            $ref: '#/components/schemas/RoleAssignment'
        organization:
          $ref: '#/components/schemas/OrganizationRef'
        status:
          $ref: '#/components/schemas/AccountStatus'
        createdAt:
          type: string
          format: date-time
        lastLoginAt:
          type: string
          format: date-time
    AccountStatus:
      type: string
      enum: [PENDING, ACTIVE, INACTIVE, SUSPENDED, RETIRED, MEMORIAL, DELETED]
    RoleAssignment:
      type: object
      properties:
        id:
          type: integer
          format: int64
        role:
          $ref: '#/components/schemas/UserRole'
        organization:
          $ref: '#/components/schemas/OrganizationRef'
        validFrom:
          type: string
          format: date
        validTo:
          type: string
          format: date
          nullable: true
    UserRole:
      type: string
      enum:
        - EMPLOYEE
        - FACULTY_SECRETARY
        - DEAN
        - RECTOR_SECRETARY
        - RECTOR
        - SYSTEM_ADMIN
        - GDPR_OFFICER
    UserUpdateRequest:
      type: object
      properties:
        firstName:
          type: string
          maxLength: 100
        lastName:
          type: string
          maxLength: 100
        notificationPreferences:
          $ref: '#/components/schemas/NotificationPreferences'
    UserSummary:
      type: object
      properties:
        id:
          type: integer
          format: int64
        email:
          type: string
          format: email
        firstName:
          type: string
        lastName:
          type: string
        organization:
          $ref: '#/components/schemas/OrganizationRef'
        status:
          $ref: '#/components/schemas/AccountStatus'
        roles:
          type: array
          description: Roles in effect today
          items:
            $ref: '#/components/schemas/RoleAssignment'
        membershipConfirmed:
          type: boolean
          description: Whether somebody has ever granted the user a role
    UserDetail:
      allOf:
        - $ref: '#/components/schemas/UserSummary'
        - type: object
          properties:
            createdAt:
              type: string
              format: date-time
            lastLoginAt:
              type: string
              format: date-time
              nullable: true
            roleHistory:
              type: array
              description: Every assignment ever made, newest first
              items:
                $ref: '#/components/schemas/RoleAssignment'
    UserPage:
      allOf:
        - $ref: '#/components/schemas/Page'
        - type: object
          properties:
            content:
              type: array
              items:
                $ref: '#/components/schemas/UserSummary'
    NotificationPreferences:
      type: object
      properties:
        emailEnabled:
          type: boolean
          default: true
        browserEnabled:
          type: boolean
          default: true

    # Organization
    OrganizationRef:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        nameUk:
          type: string
        code:
          type: string
        type:
          type: string
          enum: [UNIVERSITY, COLLEGE, FACULTY, SPECIALITY, DEPARTMENT]

    # Award
...
    Page:
      type: object
      properties:
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
        size:
          type: integer
        number:
          type: integer
        first:
          type: boolean
        last:
          type: boolean
    ProblemDetail:
      type: object
      description: RFC 7807 Problem Details
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
          format: uri
        timestamp:
          type: string
          format: date-time
        errors:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
  responses:
...
    Forbidden:
      description: Insufficient permissions
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
...
security:
  - bearerAuth: []